Create lesson

+++ next-app/src/hooks/mutations/use-update-user-avatar.ts
@@ -1,4 +1,4 @@
-import updateAvatar from "@/actions/update-avatar";
+import { updateAvatarAction } from "@/actions/avatar.action";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useSession } from "next-auth/react";
 import { toast } from "sonner";
@@ -15,10 +15,11 @@ export const useUpdateUserAvatar = () => {
   return useMutation({
     mutationFn: async (formData: FormData) => {
       const img = data?.user?.image?.split("/").pop();
-
-      await fetch(`http://localhost:8000/avatar/${img}`, {
-        method: "DELETE",
-      });
+      if (img) {
+        await fetch(`http://localhost:8000/avatar/${img}`, {
+          method: "DELETE",
+        });
+      }
 
       const res = await fetch(`http://localhost:8000/avatar`, {
         method: "POST",
@@ -27,7 +28,7 @@ export const useUpdateUserAvatar = () => {
 
       const { image_path } = await res.json();
       const imgPath = `http://upload-service:8000/avatar/${image_path}`;
-      return updateAvatar(imgPath);
+      return updateAvatarAction(imgPath);
     },
     onSuccess: (data) => {
       if (data.error) toast.error(data.error);
